% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_data.R
\name{ona_data}
\alias{ona_data}
\title{Download Form Data}
\usage{
ona_data(
  form_id,
  do_not_split_select_multiples = FALSE,
  include_images = FALSE,
  remove_group_name = TRUE,
  include_labels = FALSE,
  include_labels_only = FALSE,
  binary_select_multiples = FALSE,
  value_select_multiples = TRUE,
  show_choice_labels = TRUE,
  include_reviews = FALSE,
  language = "English (en)",
  group_delimiter = "/",
  start_date = NULL,
  end_date = NULL,
  ...
)
}
\arguments{
\item{form_id}{The form's unique identifier}

\item{do_not_split_select_multiples}{Split select multiple answers into separate columns when set to \code{FALSE}}

\item{include_images}{Include images when set to \code{TRUE}}

\item{remove_group_name}{Remove prefixed group names when set to \code{TRUE}}

\item{include_labels}{Include column labels when set to \code{TRUE}}

\item{include_labels_only}{Exclude question names from the header when set to \code{TRUE}}

\item{binary_select_multiples}{Use 0 or 1 in split select multiples as opposed to TRUE and FALSE when set to \code{TRUE}}

\item{value_select_multiples}{Use values for select multiple questions}

\item{show_choice_labels}{Use choice labels}

\item{include_reviews}{Include reviews when set to \code{TRUE}}

\item{language}{The export language, defaults to English}

\item{group_delimiter}{The character used to separate group names from column names. Defaults to '/'}

\item{start_date}{Filter for submissions on \code{start_date} or later}

\item{end_date}{Filter for submissions on \code{start_date} or earlier}

\item{...}{Arguments passed to \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\value{
Data Frame
}
\description{
A thin wrapper around \code{\link[=ona_create_form_export]{ona_create_form_export()}} that handles checking
the export status and saves it in \code{path} when the export succeeds.
}
