% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_qr_code.R
\name{ona_generate_qr_code}
\alias{ona_generate_qr_code}
\title{Generate QR Code}
\usage{
ona_generate_qr_code(
  project_id,
  username,
  password,
  autosend = "wifi_and_cellular",
  automatic_update = TRUE,
  form_update_mode = "previously_downloaded",
  periodic_form_updates_check = "every_fifteen_minutes"
)
}
\arguments{
\item{project_id}{project ID}

\item{username}{username}

\item{password}{password}

\item{autosend}{autosend mode, one of ("off", "wifi_only", "cellular_only", "wifi_and_cellular")}

\item{automatic_update}{whether to automatically update forms}

\item{form_update_mode}{form update mode, one of ("manual", "previously_downloaded", "match_exactly")}

\item{periodic_form_updates_check}{periodicity of form updates, one of ("every_fifteen_minutes", "every_one_hour", "every_six_hours", "every_24_hours")}
}
\value{
a ggplot2 plot
}
\description{
Generate QR Code
}
