% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_get_authenticated_request.R
\name{ona_get_authenticated_request}
\alias{ona_get_authenticated_request}
\title{Get an ONA API request with HTTP Basic Authentication}
\usage{
ona_get_authenticated_request(
  base_url = "https://api.ona.io/api/v1",
  ona_username = Sys.getenv("ONA_USERNAME", unset = NA),
  ona_password = Sys.getenv("ONA_PASSWORD", unset = NA)
)
}
\arguments{
\item{base_url}{ONA Base API URL}

\item{ona_username}{ONA Username, defaults to the environment variable ONA_USERNAME}

\item{ona_password}{ONA Password, defaults to the environment variable ONA_PASSWORD}
}
\value{
A request object with HTTP Basic Authentication
}
\description{
Get an ONA API request with HTTP Basic Authentication
}
\seealso{
\link[httr2:request]{httr2::request}

\link[httr2:req_auth_basic]{httr2::req_auth_basic}
}
