% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_get_project_info.R
\name{ona_get_project_info}
\alias{ona_get_project_info}
\title{Get Project Info from ONA}
\usage{
ona_get_project_info(project_id)
}
\arguments{
\item{project_id}{Project Identifier on ONA}
}
\value{
JSON response from ONA
}
\description{
Get Project Info from ONA
}
