% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_get_project_info.R
\name{ona_get_project_users}
\alias{ona_get_project_users}
\title{Get User List working on an ONA project}
\usage{
ona_get_project_users(project_id)
}
\arguments{
\item{project_id}{Project Identifier on ONA}
}
\value{
A list of ONA usernames
}
\description{
Get User List working on an ONA project
}
