% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_projects.R
\name{ona_projects}
\alias{ona_projects}
\title{Get the list of projects available to the authenticated user}
\usage{
ona_projects()
}
\value{
An HTTP response
}
\description{
Get the list of projects available to the authenticated user
}
\seealso{
\code{\link[httr2:response]{httr2::response()}}
}
