% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_qr_code.R
\name{ona_save_qr_code}
\alias{ona_save_qr_code}
\title{Save an ONA QR code to a file}
\usage{
ona_save_qr_code(qr_code, path)
}
\arguments{
\item{qr_code}{a QR code generated by \link{ona_generate_qr_code}}

\item{path}{a path to where the QR code will be saved}
}
\value{
invisibly return the provided path
}
\description{
Save an ONA QR code to a file
}
