% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_share_project.R
\name{ona_share_project}
\alias{ona_share_project}
\title{Share an ONA Project with a user and specify their role}
\usage{
ona_share_project(project_id, username, role = "dataentry")
}
\arguments{
\item{project_id}{Unique Identifier of the Project}

\item{username}{The username with whom the project will be shared}

\item{role}{One of \code{readonly}, \code{dataentry}, \code{editor}, and \code{manager}}
}
\value{
A HTTP response. On success the response will be empty with a HTTP Status code 204
}
\description{
Share an ONA Project with a user and specify their role
}
