% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_xlsform.R
\name{ona_xlsform}
\alias{ona_xlsform}
\title{Get the XLSForm Representation}
\usage{
ona_xlsform(form_id)
}
\arguments{
\item{form_id}{Form Unique Identifier}
}
\value{
A list of data frames, where the names of the list are sheet names
}
\description{
Get the XLSForm Representation
}
