% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONAR-package.R
\docType{package}
\name{ONAR-package}
\alias{ONAR}
\alias{ONAR-package}
\title{ONAR: A wrapper around the ONA API}
\description{
A wrapper around the ONA API.
}
\author{
\strong{Maintainer}: Iyed Ghedamsi \email{iyed.ghedamsi@gmail.com} [contributor]

}
\keyword{internal}
