% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsform.R
\name{read_xlsform}
\alias{read_xlsform}
\title{Read an XLSForm Excel file}
\usage{
read_xlsform(path, ...)
}
\arguments{
\item{path}{path to the XLSForm file}

\item{...}{Added for extensibility}
}
\value{
list of data frames where each sheet corresponds to a data frame
}
\description{
Read an XLSForm Excel file
}
