% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ona_register_user.R
\name{register_user}
\alias{register_user}
\title{Register a new user on ONA}
\usage{
register_user(user_name, email, name, password)
}
\arguments{
\item{user_name}{User Name}

\item{email}{Email}

\item{name}{Display Name}

\item{password}{Password}
}
\value{
a response from \code{\link[httr2:response]{httr2::response()}}
}
\description{
Register a new user on ONA
}
